change msgCount from size_t to uint32_t

+++ libgluon/include/galois/graphs/GluonSubstrate.h
@@ -1991,7 +1991,7 @@ public:
                     bufLen = net.AGG_MSG_SIZE;
                 }
                 else {
-                    bufLen -= sizeof(size_t);
+                    bufLen -= sizeof(uint32_t);
                 }
                 size_t offset = 0;
 
@@ -2036,12 +2036,12 @@ public:
             net.receiveRemoteWork(terminateFlag, fullFlag, buf, bufLen);
 
             if (!terminateFlag) { // received message
-                size_t msgCount;
+                uint32_t msgCount;
                 if (fullFlag) {
                     msgCount = net.WORK_COUNT;
                 }
                 else {
-                    msgCount = *((size_t*)(buf + bufLen - sizeof(size_t)));
+                    msgCount = *((uint32_t*)(buf + bufLen - sizeof(uint32_t)));
                 }
 
                 galois::on_each(
